update modify package script for none Windows OS

+++ src/Couchbase.Lite.Tests.NetCore/modify_packages.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
+if [ -f Couchbase.Lite.Tests.NetCore.Source.csproj ];
+then
 pushd `dirname $0`
 sed "s/[0-9].[0-9].[0-9]-b..../$1-$2/g" Couchbase.Lite.Tests.NetCore.csproj > tmp
 mv tmp Couchbase.Lite.Tests.NetCore.csproj
 sed "s/[0-9].[0-9].[0-9]-b..../$1-$2/g" Couchbase.Lite.Tests.NetCore.Source.csproj > tmp
 mv tmp Couchbase.Lite.Tests.NetCore.Source.csproj
 popd
+fi
